Use community mailer instead of pom mailer
