+++ app/mailers/community_mailer.rb
@@ -15,4 +15,23 @@ class CommunityMailer < GovukNotifyRails::Mailer
 
     mail(to: ldu.email_address)
   end
+
+  def handover_chase_email(nomis_offender_id:, offender_name:, offender_crn:, ldu_email:, prison:,
+                           start_date:, responsibility_handover_date:, pom_name:, pom_email:)
+    set_template('d7366b11-c93e-48de-824f-cb80a9778e71')
+
+    set_personalisation(
+    email: ldu_email,
+    name: offender_name,
+    crn: offender_crn,
+    noms_no: nomis_offender_id,
+    prison_name: prison,
+    start_date: start_date,
+    responsibility_handover_date: responsibility_handover_date,
+    pom_name: pom_name,
+    pom_email: pom_email
+    )
+
+    mail(to: ldu_email)
+  end
 end
