+++ app/mailers/pom_mailer.rb
@@ -151,24 +151,4 @@ class PomMailer < GovukNotifyRails::Mailer
 
     mail(to: params[:previous_pom_email])
   end
-
-  # TODO: Move this to the new mailer created by Stephen
-  def handover_chase_email(nomis_offender_id:, offender_name:, offender_crn:, ldu_email:, prison:,
-                           start_date:, responsibility_handover_date:, pom_name:, pom_email:)
-    set_template('d7366b11-c93e-48de-824f-cb80a9778e71')
-
-    set_personalisation(
-      email: ldu_email,
-      name: offender_name,
-      crn: offender_crn,
-      noms_no: nomis_offender_id,
-      prison_name: prison,
-      start_date: start_date,
-      responsibility_handover_date: responsibility_handover_date,
-      pom_name: pom_name,
-      pom_email: pom_email
-    )
-
-    mail(to: ldu_email)
-  end
 end
