+++ spec/jobs/handover_follow_up_job_spec.rb
@@ -84,7 +84,7 @@ RSpec.describe HandoverFollowUpJob, type: :job do
 
     expect { described_class.perform_now(date_of_handover) }
       .to have_enqueued_job(ActionMailer::MailDeliveryJob)
-          .with("PomMailer", "handover_chase_email", "deliver_now",
+          .with("CommunityMailer", "handover_chase_email", "deliver_now",
                 args: [{ nomis_offender_id: determinate_offender.offender_no,
                           offender_name: determinate_offender.full_name,
                           offender_crn: determinate_offender.crn,
